requirements.txt
Flask==2.1.2
dlib==19.24.0
face-recognition==1.3.0
numpy==1.22.4
tensorflow==2.9.1
opencv-python==4.5.5.64
pymongo==4.1.1
pillow==9.1.1